+++ lib/src/widgets/responsive_list_view.dart
@@ -6,34 +6,48 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_gehry/flutter_gehry.dart';
 
 /// Wrapper around [ListView].
+/// If a value is missing and is required by [ListView], the [ListView]'s default constructor's default value will be used.
 class ResponsiveListView extends StatelessWidget with ResponsiveStatelessWidgetMixin {
-  final BreakpointsData<ResponsiveListViewData> data;
+  final BreakpointsData<Axis>? scrollDirection;
+  final BreakpointsData<bool>? reverse;
+  final BreakpointsData<ScrollPhysics>? physics;
+  final BreakpointsData<bool>? shrinkWrap;
+  final BreakpointsData<EdgeInsets>? padding;
+  final BreakpointsData<DragStartBehavior>? dragStartBehavior;
+  final BreakpointsData<ScrollViewKeyboardDismissBehavior>? keyboardDismissBehavior;
+  final BreakpointsData<Clip>? clipBehavior;
 
-  /// Passed to [ListView.custom]
+  /// Passed to [ListView.custom].
   final SliverChildDelegate childrenDelegate;
 
-  /// Passed to [ListView]
+  /// Passed to [ListView].
   final ScrollController? controller;
 
-  /// Passed to [ScrollView.primary]
+  /// Passed to [ScrollView.primary].
   final bool? primary;
+
+  /// Passed to [ListView.cacheExtent].
+  final double? cacheExtent;
 
   /// Same as [ListView]
   ResponsiveListView({
     Key? key,
-    ResponsiveListViewData? xs,
-    ResponsiveListViewData? sm,
-    ResponsiveListViewData? md,
-    ResponsiveListViewData? lg,
-    ResponsiveListViewData? xl,
+    this.scrollDirection,
+    this.reverse,
     this.controller,
     this.primary,
+    this.physics,
+    this.shrinkWrap,
+    this.padding,
+    this.cacheExtent,
+    this.dragStartBehavior,
+    this.keyboardDismissBehavior,
+    this.clipBehavior,
     bool addAutomaticKeepAlives = true,
     bool addRepaintBoundaries = true,
     bool addSemanticIndexes = true,
     List<Widget> children = const [],
-  })  : data = BreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
-        childrenDelegate = SliverChildListDelegate(
+  })  : childrenDelegate = SliverChildListDelegate(
           children,
           addAutomaticKeepAlives: addAutomaticKeepAlives,
           addRepaintBoundaries: addRepaintBoundaries,
@@ -44,20 +58,23 @@ class ResponsiveListView extends StatelessWidget with ResponsiveStatelessWidgetM
   /// Same as [ListView.builder]
   ResponsiveListView.builder({
     Key? key,
-    ResponsiveListViewData? xs,
-    ResponsiveListViewData? sm,
-    ResponsiveListViewData? md,
-    ResponsiveListViewData? lg,
-    ResponsiveListViewData? xl,
+    this.scrollDirection,
+    this.reverse,
     this.controller,
     this.primary,
+    this.physics,
+    this.shrinkWrap,
+    this.padding,
+    this.cacheExtent,
+    this.dragStartBehavior,
+    this.keyboardDismissBehavior,
+    this.clipBehavior,
     required IndexedWidgetBuilder itemBuilder,
     int? itemCount,
     bool addAutomaticKeepAlives = true,
     bool addRepaintBoundaries = true,
     bool addSemanticIndexes = true,
-  })  : data = BreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
-        childrenDelegate = SliverChildBuilderDelegate(
+  })  : childrenDelegate = SliverChildBuilderDelegate(
           itemBuilder,
           childCount: itemCount,
           addAutomaticKeepAlives: addAutomaticKeepAlives,
@@ -69,21 +86,24 @@ class ResponsiveListView extends StatelessWidget with ResponsiveStatelessWidgetM
   /// Same as [ListView.separated]
   ResponsiveListView.separated({
     Key? key,
-    ResponsiveListViewData? xs,
-    ResponsiveListViewData? sm,
-    ResponsiveListViewData? md,
-    ResponsiveListViewData? lg,
-    ResponsiveListViewData? xl,
+    this.scrollDirection,
+    this.reverse,
     this.controller,
     this.primary,
+    this.physics,
+    this.shrinkWrap,
+    this.padding,
+    this.cacheExtent,
+    this.dragStartBehavior,
+    this.keyboardDismissBehavior,
+    this.clipBehavior,
     required IndexedWidgetBuilder itemBuilder,
     required IndexedWidgetBuilder separatorBuilder,
     required int itemCount,
     bool addAutomaticKeepAlives = true,
     bool addRepaintBoundaries = true,
     bool addSemanticIndexes = true,
-  })  : data = BreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
-        childrenDelegate = SliverChildBuilderDelegate(
+  })  : childrenDelegate = SliverChildBuilderDelegate(
           (context, index) {
             final int itemIndex = index ~/ 2;
             if (index.isEven) return itemBuilder(context, itemIndex);
@@ -97,15 +117,19 @@ class ResponsiveListView extends StatelessWidget with ResponsiveStatelessWidgetM
         super(key: key);
 
   /// Same as [ListView.custom]
-  ResponsiveListView.custom({
+  const ResponsiveListView.custom({
     Key? key,
-    ResponsiveListViewData? xs,
-    ResponsiveListViewData? sm,
-    ResponsiveListViewData? md,
-    ResponsiveListViewData? lg,
-    ResponsiveListViewData? xl,
+    this.scrollDirection,
+    this.reverse,
     this.controller,
     this.primary,
+    this.physics,
+    this.shrinkWrap,
+    this.padding,
+    this.cacheExtent,
+    this.dragStartBehavior,
+    this.keyboardDismissBehavior,
+    this.clipBehavior,
     required this.childrenDelegate,
     required IndexedWidgetBuilder itemBuilder,
     required IndexedWidgetBuilder separatorBuilder,
@@ -113,53 +137,24 @@ class ResponsiveListView extends StatelessWidget with ResponsiveStatelessWidgetM
     bool addAutomaticKeepAlives = true,
     bool addRepaintBoundaries = true,
     bool addSemanticIndexes = true,
-  })  : data = BreakpointsData(xs: xs, sm: sm, md: md, lg: lg, xl: xl),
-        super(key: key);
+  }) : super(key: key);
 
   @override
   Widget buildResponsive(BuildContext context, Breakpoints breakpoint) {
-    final listViewData = data.getForBreakpoint(breakpoint);
-
     return ListView.custom(
-      scrollDirection: listViewData?.scrollDirection ?? ResponsiveListViewData.initial.scrollDirection,
-      reverse: listViewData?.reverse ?? ResponsiveListViewData.initial.reverse,
+      scrollDirection: scrollDirection?.getForBreakpoint(breakpoint) ?? Axis.vertical,
+      reverse: reverse?.getForBreakpoint(breakpoint) ?? false,
       controller: controller,
       primary: primary,
-      physics: listViewData?.physics,
-      shrinkWrap: listViewData?.shrinkWrap ?? ResponsiveListViewData.initial.shrinkWrap,
-      padding: listViewData?.padding,
-      cacheExtent: listViewData?.cacheExtent,
-      dragStartBehavior: listViewData?.dragStartBehavior ?? ResponsiveListViewData.initial.dragStartBehavior,
+      physics: physics?.getForBreakpoint(breakpoint),
+      shrinkWrap: shrinkWrap?.getForBreakpoint(breakpoint) ?? false,
+      padding: padding?.getForBreakpoint(breakpoint),
+      cacheExtent: cacheExtent,
+      dragStartBehavior: dragStartBehavior?.getForBreakpoint(breakpoint) ?? DragStartBehavior.start,
       keyboardDismissBehavior:
-          listViewData?.keyboardDismissBehavior ?? ResponsiveListViewData.initial.keyboardDismissBehavior,
-      clipBehavior: listViewData?.clipBehavior ?? ResponsiveListViewData.initial.clipBehavior,
+          keyboardDismissBehavior?.getForBreakpoint(breakpoint) ?? ScrollViewKeyboardDismissBehavior.manual,
+      clipBehavior: clipBehavior?.getForBreakpoint(breakpoint) ?? Clip.hardEdge,
       childrenDelegate: childrenDelegate,
     );
   }
-}
-
-/// All of this data is provided to [ListView], with a fallback on [initial] if no data is provided for the current breakpoint
-class ResponsiveListViewData {
-  static const initial = ResponsiveListViewData();
-
-  final Axis scrollDirection;
-  final bool reverse;
-  final ScrollPhysics? physics;
-  final bool shrinkWrap;
-  final EdgeInsets? padding;
-  final double? cacheExtent;
-  final DragStartBehavior dragStartBehavior;
-  final ScrollViewKeyboardDismissBehavior keyboardDismissBehavior = ScrollViewKeyboardDismissBehavior.manual;
-  final Clip clipBehavior;
-
-  const ResponsiveListViewData({
-    this.scrollDirection = Axis.vertical,
-    this.reverse = false,
-    this.physics,
-    this.shrinkWrap = false,
-    this.padding,
-    this.cacheExtent,
-    this.dragStartBehavior = DragStartBehavior.start,
-    this.clipBehavior = Clip.hardEdge,
-  });
 }